pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bear-rael-openenv"
version = "1.0.0"
description = "Bayesian Embodied Autonomous Robotics Lab — OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"
